% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_package_yaml.R
\name{gen_package_yaml}
\alias{gen_package_yaml}
\title{Create the packages.yml for installed packages}
\usage{
gen_package_yaml()
}
\value{
A tibble
}
\description{
Create the packages.yml for installed packages
}
\examples{
\dontrun{
gen_package_yaml()
}
}
