% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cran_pkgs.R
\name{get_cran_pkgs}
\alias{get_cran_pkgs}
\title{Get available packages from CRAN}
\usage{
get_cran_pkgs(chunk_size = 200, timeout = 0.01)
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble}} of package metadata.
}
\description{
Returns a \code{\link[tibble:tibble]{tibble::tibble}} based on \code{\link[pkgsearch:cran_packages]{pkgsearch::cran_packages()}} for all R packages
returned by \code{\link[utils:available.packages]{utils::available.packages()}}
}
\examples{
get_cran_pkgs()
}
