% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cran_pkgs.R
\name{write_cran_meta}
\alias{write_cran_meta}
\title{Write (and cleanup) the CRAN package metadata}
\usage{
write_cran_meta(metadata, file_out = "cran.yml")
}
\arguments{
\item{metadata}{The metadata as returned by \code{\link[=get_cran_pkgs]{get_cran_pkgs()}}.}

\item{file_out}{\verb{[cran.yaml]} The output \code{YAML} file relative to \verb{data/packages/}}
}
\value{

}
\description{
Write (and cleanup) the CRAN package metadata
}
\examples{
\dontrun{
get_cran_pkgs() \%>\%
  write_cran_meta()
}
}
