% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cran_pkgs.R
\name{get_cran_pkgs}
\alias{get_cran_pkgs}
\title{Get available packages from CRAN}
\usage{
get_cran_pkgs(chunk_size = 200, timeout = 1)
}
\value{

}
\description{
Returns a single two-column tibble based on \code{\link[utils:available.packages]{utils::available.packages()}}.
}
\examples{
get_cran_pkgs()
}
